// src/app/07_SIR.Mantenimiento.Apps/interfaces/IndicadorCharqueador.interface.ts
export interface IndicadorCharqueador {
    linea: string;
    nombreEstacion: string;
    idEstacion: string;
    charqNum: string;
    charqueador: string;
    cortesRecibidos: number;
    cortesEnviados: number;
    kgRecibidos: number;
    kgEnviados: number;
    porcRendimiento: number;
    procRendPromedio: number;
    promedioKilosSalidaCharqueador: number;
}